+++ web/lib/components/word/Word.jsx
@@ -1,106 +1,34 @@
-import { useTranslation } from '@/lib/i18n'
+import { useState } from 'react'
 
-import { LinkToWord } from '../common/LinkToWord'
+import { ZapIcon } from '../common/Icons'
 
 import { WordMenu } from '../word-menu/WordMenu'
 
-import { VerbTenses } from './VerbTenses'
 import { Speech } from './Speech'
 import { Pronunciations } from './Pronunciations'
-import { Examples } from './Examples'
-import { Idioms } from './Idioms'
-import { FlashcardDialog } from './FlashcardDialog'
+import { WordDefinition } from './WordDefinition'
 
-function Definition({ def, index }) {
-  const { _e } = useTranslation()
-
-  const id = `def-${def.id}`
-
+function BtnCreatFlashcard({ active, onClick }) {
   return (
-    <div className="flex flex-col gap-4" id={id}>
-      <p className="font-semibold">
-        {index && (
-          <a
-            title={_e('word.copyPermalink')}
-            href={'#' + id}
-            className="mr-1 text-sm text-slate-400 hover:text-slate-600">
-            {index}.
-          </a>
-        )}
-
-        <span className="text-pink-700">{def.meaning}</span>
-      </p>
-
-      {def.text && <p className="ml-4">{def.text}</p>}
-
-      {def.examples && (
-        <div className="ml-4">
-          <Examples examples={def.examples} />
-        </div>
-      )}
-
-      {def.relatedWords && (
-        <div className="ml-4">
-          <SeeAlso words={def.relatedWords} />
-        </div>
-      )}
-    </div>
-  )
-}
-
-function SeeAlso({ words }) {
-  const { _e } = useTranslation()
-
-  if (!Array.isArray(words) || words.length === 0) return null
-
-  return (
-    <>
-      <p className="upper-first mb-2 font-semibold">
-        {_e('word.seeAlso', { count: words.length })}:
-      </p>
-
-      <div className="flex flex-wrap items-center gap-2">
-        {words.map((word) => (
-          <LinkToWord
-            key={word}
-            query={{ word }}
-            className="border-b-2 border-sky-600">
-            {word}
-          </LinkToWord>
-        ))}
-      </div>
-    </>
-  )
-}
-
-function WordDefinition({ def, tenses }) {
-  const isVerb =
-    def.partOfSpeech?.includes('động từ') || def.partOfSpeech === 'verb'
-
-  return (
-    <div className="grid gap-4 lg:grid-cols-2">
-      <div className="flex flex-col gap-4">
-        <h3 className="mt-2">
-          <span className="rounded bg-sky-100 p-2 font-bold uppercase text-sky-600">
-            {def.partOfSpeech}
-          </span>
-        </h3>
-
-        {isVerb && <VerbTenses tenses={tenses} />}
-
-        {def.definitions.map((d, index) => (
-          <Definition def={d} index={index + 1} key={d.meaning} />
-        ))}
-      </div>
-
-      <div>
-        <Idioms idioms={def.idioms} />
-      </div>
-    </div>
+    <button
+      type="button"
+      onClick={onClick}
+      className={`flex items-center gap-2 rounded-full border border-orange-200 p-2 text-orange-700 hover:border-orange-300 lg:rounded lg:px-4 ${
+        active ? 'bg-orange-200' : 'bg-orange-100'
+      } `}>
+      <span className="hidden text-sm font-semibold tracking-wide lg:inline-block">
+        Tạo flashcard
+      </span>
+      <ZapIcon />
+    </button>
   )
 }
 
 export function Word({ word: w }) {
+  const [flashcardMode, setFlashcardMode] = useState(false)
+
+  const doToggleFlashcardMode = () => setFlashcardMode((old) => !old)
+
   return (
     <div
       className="relative grid gap-2"
@@ -119,7 +47,11 @@ export function Word({ word: w }) {
       <div className="sticky inset-y-20">
         <div className="flex items-center gap-2">
           <WordMenu />
-          <FlashcardDialog />
+
+          <BtnCreatFlashcard
+            onClick={doToggleFlashcardMode}
+            active={flashcardMode}
+          />
         </div>
       </div>
 
@@ -129,6 +61,7 @@ export function Word({ word: w }) {
             key={def.partOfSpeech + i}
             def={def}
             tenses={w.tenses}
+            flashcardMode={flashcardMode}
           />
         ))}
       </div>
